+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- configure logging through `log-level` / `CMEEL_LOG_LEVEL`
+
 ## [v0.30.0] - 2023-01-31
 
 - on Apple Silicon, explicitely build for arm64